+++ src/main/java/um/tesoreria/guarani/hexagonal/guarani/requisitos/requisito/application/service/RequisitoGuaraniService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import um.tesoreria.guarani.hexagonal.guarani.requisitos.requisito.domain.model.RequisitoGuarani;
 import um.tesoreria.guarani.hexagonal.guarani.requisitos.requisito.domain.ports.in.GetAllRequisitosUseCase;
 import um.tesoreria.guarani.hexagonal.guarani.requisitos.requisito.domain.ports.in.GetRequisitoGuaraniUseCase;
+import um.tesoreria.guarani.hexagonal.guarani.requisitos.requisito.domain.ports.in.GetRequisitosByTipoUseCase;
 
 import java.util.List;
 
@@ -14,6 +15,7 @@ public class RequisitoGuaraniService {
 
     private final GetRequisitoGuaraniUseCase getRequisitoGuaraniUseCase;
     private final GetAllRequisitosUseCase getAllRequisitosUseCase;
+    private final GetRequisitosByTipoUseCase getRequisitosByTipoUseCase;
 
     public RequisitoGuarani getByRequisitoId(Integer requisitoId) {
         return getRequisitoGuaraniUseCase.getByRequisitoId(requisitoId);
@@ -21,5 +23,9 @@ public class RequisitoGuaraniService {
 
     public List<RequisitoGuarani> getAllRequisitos() {
         return getAllRequisitosUseCase.getAll();
+    }
+
+    public List<RequisitoGuarani> getByRequisitoTipo(Integer requisitoTipo) {
+        return getRequisitosByTipoUseCase.getByRequisitoTipo(requisitoTipo);
     }
 }