+++ src/main/java/um/tesoreria/guarani/hexagonal/guarani/requisitos/requisito/domain/ports/in/GetRequisitosByTipoUseCase.java
@@ -1,0 +1,9 @@
+package um.tesoreria.guarani.hexagonal.guarani.requisitos.requisito.domain.ports.in;
+
+import um.tesoreria.guarani.hexagonal.guarani.requisitos.requisito.domain.model.RequisitoGuarani;
+
+import java.util.List;
+
+public interface GetRequisitosByTipoUseCase {
+    List<RequisitoGuarani> getByRequisitoTipo(Integer requisitoTipo);
+}